admin/password_bit/deploy/flask_app/requirements.txt
docker==7.1.0
Flask==3.0.3
python-dotenv==1.0.0
Pillow==10.2.0  # For generating captcha images
